requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic==2.12.5
pydantic[email]==2.12.5
jinja2==3.1.6
weasyprint==66.0
httpx